changed dashboard page layout
