+++ frontend/src/App.css
@@ -1,25 +1,15 @@
 body {
-  margin: 0;
-  padding: 0;
   display: flex;
-  justify-content: center;
-  /* CRAZY MARGIN TODO */
   margin-top: 5em;
-  min-height: 100vh;
-  height: auto;
-  background-color: #f3f4f4;
+  background-color: fff;
+  justify-content: center;
   color: black;
   font-family: Arial, sans-serif;
   overflow-y: auto;
 }
 
-h1 {
-  text-align: center;
-  margin-bottom: 20px;
-}
-
 * {
-  padding: 0;
-  margin: 0;
   box-sizing: border-box;
+  margin: 0;
+  padding: 0;
 }
