+++ frontend/src/App.jsx
@@ -11,8 +11,8 @@ import ProblemPage from "./pages/ProblemPage.jsx";
 import ErrorPage from "./pages/ErrorPage.tsx";
 import RootPage from "./pages/RootPage.jsx";
 import HomeRootPage from "./pages/HomeRootPage.jsx";
-import DashboardPage from "./pages/DashboardPage.tsx";
-import { loader as topicsLoader } from "./pages/DashboardPage.tsx";
+import DashboardPage from "./pages/Dashboard/DashboardPage.tsx";
+import { loader as topicsLoader } from "./pages/Dashboard/DashboardPage.tsx";
 import { loader as problemLoader } from "./pages/ProblemPage.jsx";
 import { loader as contentLoader } from "./pages/ContentPage.jsx";
 import { action as signupAction } from "./pages/SignupPage.jsx";
