+++ frontend/src/pages/RootPage.jsx
@@ -1,18 +1,16 @@
 import React from "react";
-import { Outlet } from 'react-router-dom'; 
+import { Outlet } from "react-router-dom";
 import PrimaryHeader from "../components/headers/PrimaryHeader/PrimaryHeader";
 
 const RootPage = () => {
-
-    return (
-        <>  
-            <PrimaryHeader />
-            <main> 
-                <Outlet />
-            </main>
-        </>
-
-    );
+  return (
+    <>
+      <PrimaryHeader />
+      <main>
+        <Outlet />
+      </main>
+    </>
+  );
 };
 
 export default RootPage;
